Fix PSR2 and rename files
