+++ tests/php/Centreon/Domain/ConfigurationLoader/child1.yaml
