+++ tests/php/Centreon/Domain/ConfigurationLoader/child3.yaml
