+++ tests/php/Centreon/Domain/ConfigurationLoader/conf.d/child2.yaml
