+++ tests/php/Centreon/Domain/ConfigurationLoader/conf.d/child3.yaml
