+++ tests/php/Centreon/Domain/ConfigurationLoader/conf.d/child5.yaml
