+++ tests/php/Centreon/Domain/ConfigurationLoader/root_file.yaml
