+++ tests/php/Centreon/Domain/ConfigurationLoader/root_file_with_loop.yaml
